chore: Move raw themes to a `constants.ts` file
