+++ mobile/src/modules/theme/useTheme.ts
@@ -1,84 +1,11 @@
-import { isSystemDarkMode } from "@missingcore/native-utils";
 import { useMemo } from "react";
 import { useColorScheme } from "react-native";
 
 import { usePreferenceStore } from "~/stores/Preference/store";
 
-import { Colors } from "~/constants/Styles";
 import type { AppColor } from "~/lib/style";
 import { isHexColor } from "~/lib/style";
-
-//#region Theme
-export const FixedTheme = {
-  primary: Colors.nRed50,
-  onPrimary: Colors.neutral100,
-  onPrimaryVariant: Colors.neutral90,
-
-  secondary: Colors.yellow50,
-  onSecondary: Colors.neutral0,
-  onSecondaryVariant: Colors.neutral30,
-
-  error: Colors.red40,
-  onError: Colors.neutral100,
-  onErrorVariant: Colors.neutral90,
-} as const;
-
-export const Themes = {
-  light: {
-    scheme: "light",
-    ...FixedTheme,
-    primaryDim: Colors.nRed45,
-    secondaryDim: Colors.yellow47,
-    errorDim: Colors.red37,
-
-    surfaceDim: Colors.neutral80,
-    surface: Colors.neutral95,
-    surfaceBright: Colors.neutral95,
-
-    surfaceContainerLowest: Colors.neutral100,
-    surfaceContainerLow: Colors.neutral98,
-    surfaceContainer: Colors.neutral92,
-    surfaceContainerHigh: Colors.neutral85,
-    surfaceContainerHighest: Colors.neutral83,
-
-    onSurface: Colors.neutral0,
-    onSurfaceVariant: Colors.neutral40,
-    outline: Colors.neutral70,
-    outlineVariant: Colors.neutral85,
-
-    inverseSurface: Colors.neutral0,
-    inverseOnSurface: Colors.neutral100,
-  },
-  dark: {
-    scheme: "dark",
-    ...FixedTheme,
-    primaryDim: Colors.nRed40,
-    secondaryDim: Colors.yellow45,
-    errorDim: Colors.red35,
-
-    surfaceDim: Colors.neutral0,
-    surface: Colors.neutral0,
-    surfaceBright: Colors.neutral7,
-
-    surfaceContainerLowest: Colors.neutral10,
-    surfaceContainerLow: Colors.neutral13,
-    surfaceContainer: Colors.neutral15,
-    surfaceContainerHigh: Colors.neutral20,
-    surfaceContainerHighest: Colors.neutral23,
-
-    onSurface: Colors.neutral100,
-    onSurfaceVariant: Colors.neutral70,
-    outline: Colors.neutral40,
-    outlineVariant: Colors.neutral20,
-
-    inverseSurface: Colors.neutral100,
-    inverseOnSurface: Colors.neutral0,
-  },
-} as const;
-
-/** Returns theme colors based on the system theme on app launch. */
-export const SystemTheme = Themes[isSystemDarkMode ? "dark" : "light"];
-//#endregion
+import { Themes } from "./constants";
 
 //#region Hooks
 /** Returns if we're using light or dark theme. */
